Revert "Feature/option support (#19)"
This reverts commit c4d0fca1459325793bf9b218c64ba76dbf5def0f.
