+++ README.md
@@ -1,34 +1,9 @@
 [![Build Status](https://travis-ci.com/Aaron-K-T-Berry/generator-ansible-molecule.svg?branch=master)](https://travis-ci.com/Aaron-K-T-Berry/generator-ansible-molecule)
 [![Coverage Status](https://coveralls.io/repos/github/Aaron-K-T-Berry/generator-ansible-molecule/badge.svg?branch=master)](https://coveralls.io/github/Aaron-K-T-Berry/generator-ansible-molecule?branch=master)
 [![Known Vulnerabilities](https://snyk.io/test/github/Aaron-K-T-Berry/generator-ansible-molecule/badge.svg)](https://snyk.io/test/github/Aaron-K-T-Berry/generator-ansible-molecule)
-# Generator-Ansible-Molecule
-Yeoman generator to create ansible roles with molecule testing framework including optional sample ci cd configs.
+# generator-ansible-molecule
 
-### Installation
-``` 
-npm i generator-ansible-molecule
-```
-
-### Running
-```
-yo ansible-molecule
-```
-
-And follow the prompts to create a new ansible role.
-
-### Support
-Currently this generator is being tested against the following node versions.
-- Node v11
-- Node v10
-- Node v8
-
-### Contributing
-Feel free to post issues, and create prs with an changes, fixes or enhancements.
-
-<!-- TODO [OPTS] Add option to specify path of install -->
-<!-- TODO [CICD] Add more ci cd options aim for all free ones on github marketplace -->
-<!-- TODO [MOLECULE] Add more molecule driver opts -->
-<!-- TODO [MOLECULE] Cleanup optional molecule files -->
-<!-- TODO [TESTS] Split up tests more logically -->
-<!-- TODO [REPO] Add semantic release to ci pipeline -->
-<!-- TODO [DRIVER] Add prompts for different driver opts -->
+# Todo
+- Add support for other ci cd platforms like travis
+- Add option to specify path of install
+- Print out all options after yo message
